refactor(dashboard): extract Court agents_demo (F-B-202 PR-10/10 — closes F-B-202)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-10 of 10. base=main. **Closes F-B-202.**

Final extraction: the agents registry list + ADR-020 principal-type
demo views (users / workloads / resources / federation). 5 routes
moved into one new sibling module.

  app/dashboard/agents_demo_routes.py (new, 137 LOC, 5 routes):
    GET  /dashboard/agents          agent registry list (with binding + ws)
    GET  /dashboard/users           ADR-020 user principals (demo)
    GET  /dashboard/workloads       ADR-020 workload principals (demo)
    GET  /dashboard/resources       ADR-020 resource principals (demo)
    GET  /dashboard/federation      federation peers + Court status (demo)

  app/dashboard/router.py (2071 → 1980 LOC, -91 net on main; full
    F-B-202 trajectory below):
    - Sections 11 (Agents) + 12 (Users/Workloads/Resources/Federation)
      removed; one breadcrumb comment at the old position.
    - One new include_router call.

Note on theme:
  - ``agents_list`` and the four ADR-020 demo views share the
    ``_demo_cast`` import (``agent_extras`` for the agents list
    enrollment_method / automation_type extras, ``users_cast`` /
    ``workloads_cast`` / ``resources_cast`` / ``peers_cast`` for the
    rest). Bundled in one module so the future ``_demo_cast`` deletion
    pass (when /v1/admin/users + /v1/admin/workloads ship) touches one
    file, not five.

Behavioural delta: zero. All five routes are read-only — no CSRF /
sealed gate touched.

Pattern follows the rest of Sprint 2: ``APIRouter(tags=[...])`` with
no prefix, bare paths, outer prefix injected by include_router.

Tests:
  - pytest tests/test_dashboard.py tests/test_dashboard_principals.py
    tests/test_dashboard_sealed_org.py tests/test_dashboard_agents_reach.py
    — 84/84
  - pytest tests/test_dashboard*.py tests/test_enrollment.py
    tests/test_onboarding.py tests/test_policy*.py tests/test_audit*.py
    tests/test_rfq*.py — 460/460 (1 pre-existing env-fail deselected:
    test_badge_approvals_empty_when_plugin_unavailable requires
    cullis_enterprise NOT installed; passes clean in CI).

```
Pre-Sprint 2 : 67 routes, 3125 LOC
Post-PR-1 #839: 67 routes, 2930 LOC   merged   helpers + template_env
Post-PR-2 #841: 60 routes, 2614 LOC   merged   auth (7 routes)
Post-PR-3 #842: 57 routes, 2528 LOC   merged   admin_settings (3 routes)
Post-PR-4 #847: 50 routes, 2372 LOC   merged   badges + sse (7 routes)
Post-PR-5 #848: 39 routes, 2071 LOC   merged   orgs (11 routes)
Post-PR-6 #849: open                          org_onboard + org_seal (7 routes)
Post-PR-7 #850: open                          agents lifecycle + credentials (12 routes)
Post-PR-8 #851: open                          policies + bindings (8 routes)
Post-PR-9 #852: open                          sessions + audit + rfq (6 routes)
Post-PR-10 (this, before #849-852 merged): 34 routes, 1980 LOC
```

After all PR-6 → PR-10 land sequentially:
  ~3 inline routes (overview + breadcrumbs only)
  ~110 LOC aggregator wiring 14 sub-routers via include_router

Conflict warning: PR-6 #849 + PR-7 #850 + PR-8 #851 + PR-9 #852 are
still open. This PR's sections (11+12) have zero line overlap with
their sections — rebase clean post-merge in any order. Only contention
is the include_router stack at top of router.py (1-line nest, trivial
to resolve).

Closes F-B-202.
