+++ app/dashboard/router.py
@@ -134,6 +134,12 @@ router.include_router(_sessions_routes.router)
 router.include_router(_audit_routes.router)
 router.include_router(_rfq_routes.router)
 
+# F-B-202 PR-10 (final): include agents_demo (agents list + ADR-020
+# principal-type demo views: users/workloads/resources/federation).
+# 5 routes — closes F-B-202.
+from app.dashboard import agents_demo_routes as _agents_demo_routes  # noqa: E402
+router.include_router(_agents_demo_routes.router)
+
 
 # ─────────────────────────────────────────────────────────────────────────────
 # Login / Logout
@@ -256,106 +262,9 @@ async def overview(request: Request, db: AsyncSession = Depends(get_db)):
 # since F-B-202 PR-5.
 
 
-# ─────────────────────────────────────────────────────────────────────────────
-# Agents
-# ─────────────────────────────────────────────────────────────────────────────
-
-@router.get("/agents", response_class=HTMLResponse)
-async def agents_list(request: Request, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-
-    q = select(AgentRecord).order_by(AgentRecord.org_id, AgentRecord.agent_id)
-    agents = (await db.execute(q)).scalars().all()
-
-    binding_statuses = {}
-    binding_q = select(BindingRecord.agent_id, BindingRecord.status).order_by(BindingRecord.id.desc())
-    for row in (await db.execute(binding_q)).all():
-        if row[0] not in binding_statuses:
-            binding_statuses[row[0]] = row[1]
-
-    agent_list = []
-    for agent in agents:
-        extras = _demo_cast.agent_extras(agent.agent_id)
-        agent_list.append({
-            "agent_id": agent.agent_id,
-            "org_id": agent.org_id,
-            "display_name": agent.display_name,
-            "is_active": agent.is_active,
-            "capabilities": agent.capabilities,
-            "binding_status": binding_statuses.get(agent.agent_id),
-            "ws_connected": ws_manager.is_connected(agent.agent_id),
-            "cert_thumbprint": agent.cert_thumbprint,
-            "enrollment_method": extras.get("enrollment_method"),
-            "automation_type": extras.get("automation_type"),
-        })
-
-    return templates.TemplateResponse("agents.html",
-        _ctx(request, session, active="agents", agents=agent_list)
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Users / Workloads / Resources / Federation (ADR-020 principal-type sections)
-#
-# Phase 1 of the SPA rework ships these views with hardcoded demo data
-# from imp/insurance-demo-spec.md. The matching admin REST endpoints
-# (/v1/admin/users, /v1/admin/workloads) are owned by the backend
-# session and land separately. Once they are live, swap the
-# ``_demo_cast.*`` calls below for httpx calls to those routes and
-# delete app/dashboard/_demo_cast.py.
-# ─────────────────────────────────────────────────────────────────────────────
-
-@router.get("/users", response_class=HTMLResponse)
-async def users_list(request: Request):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    return templates.TemplateResponse("users.html", _ctx(
-        request, session, active="users",
-        users=_demo_cast.users_cast(),
-        endpoint_ready=False,
-    ))
-
-
-@router.get("/workloads", response_class=HTMLResponse)
-async def workloads_list(request: Request):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    return templates.TemplateResponse("workloads.html", _ctx(
-        request, session, active="workloads",
-        workloads=_demo_cast.workloads_cast(),
-        endpoint_ready=False,
-    ))
-
-
-@router.get("/resources", response_class=HTMLResponse)
-async def resources_list(request: Request):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    return templates.TemplateResponse("resources.html", _ctx(
-        request, session, active="resources",
-        resources=_demo_cast.resources_cast(),
-    ))
-
-
-@router.get("/federation", response_class=HTMLResponse)
-async def federation_view(request: Request):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    return templates.TemplateResponse("federation.html", _ctx(
-        request, session, active="federation",
-        peers=_demo_cast.peers_cast(),
-        court_status="Online",
-        court_endpoint="https://court.cullis.test",
-        court_audit_status="Healthy",
-        court_last_anchor="active",
-        court_last_anchor_iso=None,
-    ))
+# Agents list (/dashboard/agents) and ADR-020 principal-type demo views
+# (/dashboard/users, /workloads, /resources, /federation) moved to
+# ``app/dashboard/agents_demo_routes.py`` since F-B-202 PR-10.
 
 
 # Sessions list (/dashboard/sessions) moved to
